пересортировка по пакетам
