+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/additional/DataPath.java
@@ -1,0 +1,39 @@
+/*
+ * This file is a part of MDClasses.
+ *
+ * Copyright © 2019 - 2020
+ * Tymko Oleg <olegtymko@yandex.ru>, Maximov Valery <maximovvalery@gmail.com> and contributors
+ *
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * MDClasses is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * MDClasses is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with MDClasses.
+ */
+package com.github._1c_syntax.mdclasses.mdo.additional;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * Модель пути к данным. Используется в элементах формы
+ */
+@Data
+@NoArgsConstructor
+public class DataPath {
+  public static final DataPath EMPTY = new DataPath();
+  private String segment = "";
+
+  public DataPath(String segment) {
+    this.setSegment(segment);
+  }
+}
