+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/form/FormData.java
@@ -24,7 +24,7 @@ package com.github._1c_syntax.mdclasses.mdo.form;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerChildItems;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerForm;
 import com.github._1c_syntax.mdclasses.mdo.wrapper.form.DesignerFormItem;
-import com.github._1c_syntax.mdclasses.metadata.additional.DataPath;
+import com.github._1c_syntax.mdclasses.mdo.additional.DataPath;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamImplicit;
 import lombok.Data;
