+++ src/test/java/com/github/_1c_syntax/mdclasses/mdo/AbstractMDOTest.java
@@ -23,8 +23,8 @@ package com.github._1c_syntax.mdclasses.mdo;
 
 import com.github._1c_syntax.mdclasses.metadata.additional.AttributeType;
 import com.github._1c_syntax.mdclasses.metadata.additional.ConfigurationSource;
-import com.github._1c_syntax.mdclasses.metadata.additional.MDOModule;
-import com.github._1c_syntax.mdclasses.metadata.additional.MDOReference;
+import com.github._1c_syntax.mdclasses.mdo.additional.MDOModule;
+import com.github._1c_syntax.mdclasses.mdo.additional.MDOReference;
 import com.github._1c_syntax.mdclasses.metadata.additional.MDOType;
 import com.github._1c_syntax.mdclasses.metadata.additional.ModuleType;
 import com.github._1c_syntax.mdclasses.metadata.additional.ObjectBelonging;
