+++ src/test/java/com/github/_1c_syntax/mdclasses/mdo/SubsystemTest.java
@@ -21,7 +21,7 @@
  */
 package com.github._1c_syntax.mdclasses.mdo;
 
-import com.github._1c_syntax.mdclasses.metadata.additional.MDOReference;
+import com.github._1c_syntax.mdclasses.mdo.additional.MDOReference;
 import com.github._1c_syntax.mdclasses.metadata.additional.MDOType;
 import io.vavr.control.Either;
 import org.junit.jupiter.api.Test;
